+++ services/analyzer/sequelize-tables-analyzer.js
@@ -301,7 +301,10 @@ async function createTableSchema({
   };
 }
 
-function detectAndFixAliasConflicts(wholeSchema) {
+// NOTICE: This detects two generated fields (regular or reference's alias) with the same name
+//         and rename reference's alias as `Linked${collectionReferenced}` to prevent Sequelize
+//         from crashing at startup.
+function fixAliasConflicts(wholeSchema) {
   const tablesName = Object.keys(wholeSchema);
 
   if (!tablesName.length) { return; }
@@ -314,7 +317,7 @@ function detectAndFixAliasConflicts(wholeSchema) {
 
       table.references.forEach((reference, index) => {
         if (fieldNames.includes(reference.as)) {
-          table.references[index].as = `${reference.as}Relationship`;
+          table.references[index].as = `linked${_.upperFirst(reference.as)}`;
         }
       });
     }
@@ -377,7 +380,7 @@ async function analyzeSequelizeTables(databaseConnection, config, allowWarning) 
     });
   }
 
-  detectAndFixAliasConflicts(schemaAllTables);
+  fixAliasConflicts(schemaAllTables);
 
   return schemaAllTables;
 }